requirements.txt
fastapi
uvicorn
DrissionPage
playwright
rich
Pillow>=11.2.1